requirements.txt
django-push-notifications
django
gunicorn
python-dotenv
celery
redis
